fix: address theme implementation review feedback
Harden theming by completing color decoupling across all UI surfaces,
adding hex color validation, unit tests, and safe deep-copy for theme
preview revert. Remove dead code, eliminate redundant resolvedPalette
struct, split built-in themes into separate file, and respect
XDG_CONFIG_HOME for custom theme discovery.

Key changes:
- Replace all remaining hardcoded charmtone references in api_key_input,
  session.go, and diffview/style.go with themed palette colors
- Add regex hex color validation (#RGB / #RRGGBB) for required and
  optional fields with clear error messages in sorted field order
- Remove resolvedPalette struct; NewStyles() calls hexColor() inline
- Split 22 built-in palette definitions into themes_builtin.go
- Derive diff color fallbacks from theme palette via blendHex() instead
  of hardcoded dark-only defaults
- Add Styles.Clone() with JSON deep-copy for ansi.StyleConfig to fix
  shallow-copy pointer aliasing during theme preview revert
- Add StylesFromConfig() helper; wire theme-aware styles into app.go,
  run.go, and session.go
- Remove dead ConfigStore.SetTheme() method
- Respect XDG_CONFIG_HOME in DiscoverThemes(); log warnings via slog
  for invalid/shadowed custom theme files
- Add 31 unit tests covering ParseTheme, LoadTheme, DiscoverThemes,
  hex validation, diff defaults, round-trip, XDG, and all builtins
